use gates and policy in project

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 
 class PermissionController extends Controller
@@ -14,6 +15,7 @@ class PermissionController extends Controller
         return view('admin.permissions.index', compact('permissions'));
     }
     public function store(Request $request){
+        $this->authorize('create', Permission::class);
         $request->validate([
             'name' => 'required|min:3|max:50',
         ]);
@@ -30,7 +32,12 @@ class PermissionController extends Controller
         $permission = Permission::findOrFail($id);
         return view('admin.permissions.edit', ['permission' => $permission]);
     }
-    public function update(Request $request, $id){
+    public function update(Request $request, Permission $permission, $id){
+
+        $this->authorize('update', $permission);
+        // if (! Gate::allows('isAdmin')) {
+        //     abort(403);
+        // }
         try{
             $permission = Permission::findOrFail($id);
             $permission->name = $request->name;
@@ -42,11 +49,14 @@ class PermissionController extends Controller
             return redirect()->back();
         }
     }
-    public function destroy(Request $request){
+    public function destroy(Permission $permission, Request $request){
+       
         try{
             $id = $request->id;
             $permission = Permission::where('id', $id)->first();
+            $this->authorize('delete', $permission);
             $permission->delete();
+            
         }catch(\Exception $e){
             echo $e->getMessage();
         }
